% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGridSpyData.R
\name{checkDates}
\alias{checkDates}
\title{Check date fomrats}
\usage{
checkDates(dt)
}
\arguments{
\item{dt}{the data table to use & return}
}
\description{
\code{checkDates} takes character column in a data.table which is thought to be a date and tries to work out what format the date is in. 
Returns the best guess as a new column dateFormat in the data.table. If can't guess, sets dateFormat to 'ambiguous'.
Assumes the character column is called 'date_char' - I suppose this ought to be parameterised but...
Requires data.table as it uses tstrsplit
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
