% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGridSpyFiles.R
\name{list1mGridSpyFiles}
\alias{list1mGridSpyFiles}
\title{List all available 1 minute grid spy data files from a given path}
\usage{
list1mGridSpyFiles(fpath, pattern)
}
\arguments{
\item{fpath}{the name of the directory where the function should look}

\item{pattern}{a pattern to match. Use the pattern to filter e.g. 1m (xx_at1.csv) from 30s () files}
}
\description{
\code{list1mGridSpyFiles}. Lists all available 1 minute grid spy files from data repository. The functions, like most others in this package will fail
   if you do not have the data in a folder/directory corresponding to 'fpath'. Code borrows extensively from similar SAVE project function.
   This could take quite a long time if you have a lot of files. The function returns a data table with 2 columns: hhID (household id) and fullPath (full path to each file)
   We need the hhID as we cannot just concatinate all the files since the column headings (circuit labels) vary.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
