% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGridSpyData.R
\name{process1minGridSpyFileList}
\alias{process1minGridSpyFileList}
\title{Process list of 1 min grid spy data files}
\usage{
process1minGridSpyFileList(dt)
}
\arguments{
\item{dt}{the file list as a data.table}
}
\value{
Retruns dt but with new columns for the meta-data
}
\description{
\code{process1minGridSpyFileList} takes a list of 1 minute data files and processes them to extract metadata and check date formats.
Returns the file list as an expanded data table.
}
\section{Requires}{

data.table for tstrsplit

lubridate for munching dates

dplyr for select/contains

outPath: place to save interim file list

fListInterim: name of interim file list
}

\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
